requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Flask-Mail==0.9.1
opencv-python==4.8.1.78
Pillow==10.0.1